app_servers/jobs/app.py
import os

from flask import Flask
from dotenv import load_dotenv
from flask_sqlalchemy import SQLAlchemy
from flask_migrate import Migrate

app = Flask(__name__)
env_config = os.getenv('APP_SETTINGS', 'config.DevelopmentConfig')
app.config.from_object(env_config)
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
db = SQLAlchemy(app)
migrate = Migrate(app, db)
load_dotenv()

from models import Repos


@app.route('/')
def home():
    """
    Route that returns the status of each job
    """
    return 'status: <none>'


if __name__ == '__main__':
    app.run()