app_servers/jobs/requirements.txt
alembic==1.15.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==3.4.1
requests==2.32.3
rq==1.2.2
setuptools==76.0.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
